Allied some reformating.
