+++ src/jace/util/jax_helper.py
@@ -73,7 +73,10 @@ class JaCeVar:
     def __hash__(self) -> int:
         return id(self)
 
-    def __eq__(self, other: Any) -> bool:
+    def __eq__(
+        self,
+        other: Any,
+    ) -> bool:
         if not isinstance(other, JaCeVar):
             return NotImplemented
         return id(self) == id(other)
@@ -99,7 +102,9 @@ def get_jax_var_name(jax_var: jax_core.Atom | JaCeVar) -> str:
             )
 
 
-def get_jax_var_shape(jax_var: jax_core.Atom | JaCeVar) -> tuple[int | dace.symbol | str, ...]:
+def get_jax_var_shape(
+    jax_var: jax_core.Atom | JaCeVar,
+) -> tuple[int | dace.symbol | str, ...]:
     """Returns the shape of `jax_var`."""
     match jax_var:
         case jax_core.Var() | jax_core.Literal():
@@ -112,7 +117,9 @@ def get_jax_var_shape(jax_var: jax_core.Atom | JaCeVar) -> tuple[int | dace.symb
             raise TypeError(f"'get_jax_var_shape()` is not implemented for '{type(jax_var)}'.")
 
 
-def get_jax_var_dtype(jax_var: jax_core.Atom | JaCeVar) -> dace.typeclass:
+def get_jax_var_dtype(
+    jax_var: jax_core.Atom | JaCeVar,
+) -> dace.typeclass:
     """Returns the DaCe equivalent of `jax_var`s datatype."""
     match jax_var:
         case jax_core.Var() | jax_core.Literal():
@@ -140,7 +147,9 @@ def is_tracing_ongoing(
     return any(isinstance(x, jax_core.Tracer) for x in itertools.chain(args, kwargs.values()))
 
 
-def translate_dtype(dtype: Any) -> dace.typeclass:
+def translate_dtype(
+    dtype: Any,
+) -> dace.typeclass:
     """Turns a Jax datatype into a DaCe datatype."""
     if dtype is None:
         raise NotImplementedError  # Handling a special case in DaCe.
@@ -201,7 +210,9 @@ def propose_jax_name(
     return jax_name
 
 
-def get_jax_literal_value(lit: jax_core.Atom) -> bool | float | int | np.generic:
+def get_jax_literal_value(
+    lit: jax_core.Atom,
+) -> bool | float | int | np.generic:
     """Returns the value a literal is wrapping.
 
     The function guarantees to return a scalar value.
